Colocando token em todas as rotas

+++ app/controllers/adm_controller.js
@@ -12,6 +12,7 @@ const AdmController = {
     }
   },
   create: (req, res, next) => {
+  if(req.headers.token == TOKEN){
     const adm = new Adm({ nome: req.body.nome, senha: req.body.senha, email: req.body.email });
     adm.save(error => {
       if(error){
@@ -21,9 +22,13 @@ const AdmController = {
       
       res.status(201).send({});
     });
+  }
+  else{
+    res.status.send({error: `Acesso negado a API`})
+  }
   },
   change: async(req, res, next) => {
-    console.log(req.params.adm_id)
+  if(req.headers.token == TOKEN){
     try{
       await Adm.findOneAndUpdate({_id: req.params.adm_id}, {nome: req.body.nome, senha: req.body.senha, email: req.body.email})
       res.status(204).send(`Alterado com o id ${req.params.adm_id}`)
@@ -31,8 +36,14 @@ const AdmController = {
     catch(err){
       res.status(401).send(`Erro: ${err}`)
     }
-  },
+  }
+  else{
+    res.status(401).send({error:`Acesso negado a API`})
+  }
+  }
+  ,
   delete: async(req, res, next) => {
+  if(req.headers.token == TOKEN){
     try{
       await Adm.findByIdAndDelete(req.params.adm_id)
       res.status(204).send({});
@@ -41,6 +52,10 @@ const AdmController = {
     catch(err){
       res.status(401).send({})
     }
+  }
+  else{
+    res.status.send({error: "Acesso negado a API"})
+  }
   }
 }
 module.exports = AdmController;
